requirements.txt
streamlit>=1.29.0
pandas>=2.1.1
matplotlib>=3.8.2
seaborn>=0.13.1
pillow>=10.1.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-core>=0.1.10
langchain-community>=0.0.20
langgraph>=0.0.20
langchain-google-genai>=0.0.5
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
psycopg2-binary>=2.9.9
sqlalchemy==2.0.25
numpy==1.24.4
google-generativeai>=0.3.1
sqlparse>=0.5.0 